Refactor python scripts and templates
