+++ dashboard/generate/definitions.py
@@ -1,0 +1,6 @@
+from pathlib import Path
+
+
+TEMPLATES_PATH = Path(__file__).parents[1] / "templates"
+PULL_REQUESTS_TEMPLATE = "pull_requests.html.template"
+CONSTANTS_TEMPLATE = "constants.tsx.template"
