+++ dashboard/generate/generate_constants.py
@@ -6,28 +6,34 @@ from pathlib import Path
 import duckdb
 from jinja2 import Environment, FileSystemLoader
 
+from definitions import CONSTANTS_TEMPLATE, TEMPLATES_PATH
 
-def generate_html(duckdb_database: str | Path, output_file: str | Path) -> None:
-    """Generate the HTML page that summarizes Xebia's Github contributions.
+
+def generate_constants(duckdb_database: str | Path, output_file: str | Path) -> None:
+    """Generate the constants page for the dashboard.
 
     Parameters
     ----------
     duckdb_database : str | Path
         The path to the Duckdb database
-    output_directory : str | Path
-        The directory to output the webpage into
+    output_file : str | Path
+        The file to output
     """
     connection = duckdb.connect(str(duckdb_database), read_only=True)
 
     pull_requests_table = "main_consumers_xebia.consm_xebia_pull_requests"
     pull_requests = connection.table(pull_requests_table).to_df()
 
-    unique_authors = pull_requests["author"].unique().tolist()
+    unique_authors = pull_requests["author"].nunique()
+    unique_pull_requests = pull_requests["url"].nunique()
+    unique_projects = pull_requests["full_repository_name"].nunique()
 
-    env = Environment(loader=FileSystemLoader(Path(__file__).parent.as_posix()))
-    template = env.get_template("pull_requests_template.html")
+    env = Environment(loader=FileSystemLoader(TEMPLATES_PATH))
+    template = env.get_template(CONSTANTS_TEMPLATE)
     html = template.render(
-        authors=unique_authors, pull_requests=pull_requests.to_records()
+        unique_authors=unique_authors,
+        unique_pull_requests=unique_pull_requests,
+        unique_projects=unique_projects,
     )
 
     output_path = Path(output_file)
@@ -37,7 +43,7 @@ def generate_html(duckdb_database: str | Path, output_file: str | Path) -> None:
 
 
 if __name__ == "__main__":
-    parser = argparse.ArgumentParser("Generate Xebia pull request website")
+    parser = argparse.ArgumentParser("Generate constants")
 
     parser.add_argument("duckdb_database", type=Path)
     parser.add_argument("output_file", type=Path)
@@ -47,4 +53,4 @@ if __name__ == "__main__":
     if not args.duckdb_database.is_file():
         raise ValueError(f"Database does not exists: {args.duckdb_database}")
 
-    generate_html(args.duckdb_database, args.output_file)
+    generate_constants(args.duckdb_database, args.output_file)
