+++ dashboard/generate/generate_html.py
@@ -6,36 +6,30 @@ from pathlib import Path
 import duckdb
 from jinja2 import Environment, FileSystemLoader
 
-
-TEMPLATES_PATH = Path(__file__).parent / "templates"
-CONSTANTS_TEMPLATE = "constants.tsx.template"
+from definitions import TEMPLATES_PATH, PULL_REQUESTS_TEMPLATE
 
 
-def generate_constants(duckdb_database: str | Path, output_file: str | Path) -> None:
-    """Generate the constants page for the dashboard.
+def generate_html(duckdb_database: str | Path, output_file: str | Path) -> None:
+    """Generate the HTML page that summarizes Xebia's Github contributions.
 
     Parameters
     ----------
     duckdb_database : str | Path
         The path to the Duckdb database
-    output_file : str | Path
-        The file to output
+    output_directory : str | Path
+        The directory to output the webpage into
     """
     connection = duckdb.connect(str(duckdb_database), read_only=True)
 
     pull_requests_table = "main_consumers_xebia.consm_xebia_pull_requests"
     pull_requests = connection.table(pull_requests_table).to_df()
 
-    unique_authors = pull_requests["author"].nunique()
-    unique_pull_requests = pull_requests["url"].nunique()
-    unique_projects = pull_requests["full_repository_name"].nunique()
+    unique_authors = pull_requests["author"].unique().tolist()
 
     env = Environment(loader=FileSystemLoader(TEMPLATES_PATH))
-    template = env.get_template(CONSTANTS_TEMPLATE)
+    template = env.get_template(PULL_REQUESTS_TEMPLATE)
     html = template.render(
-        unique_authors=unique_authors,
-        unique_pull_requests=unique_pull_requests,
-        unique_projects=unique_projects,
+        authors=unique_authors, pull_requests=pull_requests.to_records()
     )
 
     output_path = Path(output_file)
@@ -45,7 +39,7 @@ def generate_constants(duckdb_database: str | Path, output_file: str | Path) -> 
 
 
 if __name__ == "__main__":
-    parser = argparse.ArgumentParser("Generate constants")
+    parser = argparse.ArgumentParser("Generate Xebia pull request website")
 
     parser.add_argument("duckdb_database", type=Path)
     parser.add_argument("output_file", type=Path)
@@ -55,4 +49,4 @@ if __name__ == "__main__":
     if not args.duckdb_database.is_file():
         raise ValueError(f"Database does not exists: {args.duckdb_database}")
 
-    generate_constants(args.duckdb_database, args.output_file)
+    generate_html(args.duckdb_database, args.output_file)
