+++ dashboard/templates/pull_requests.html.template
